# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MyDesk",
    "version": "1.0"
  },
  "paths": {
...
    "/api/tiles": {
      "get": {
        "tags": [
          "Tile"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TileOutput"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TileOutput"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TileOutput"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TileInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TileInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TileInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TileOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TileOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TileOutput"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tiles/{tileId}": {
      "put": {
        "tags": [
          "Tile"
        ],
        "parameters": [
          {
            "name": "tileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TileInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TileInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TileInput"
              }
            }
          }
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TileOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TileOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TileOutput"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tile"
        ],
        "parameters": [
          {
            "name": "tileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/tiles/{zoneId}": {
      "get": {
        "tags": [
          "Tile"
        ],
        "parameters": [
          {
            "name": "zoneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TileOutput"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TileOutput"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TileOutput"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "TileInput": {
        "type": "object",
        "properties": {
          "zone_Id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "$ref": "#/components/schemas/TileType"
          },
          "x_Position": {
            "type": "integer",
            "format": "int32"
          },
          "y_Position": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TileOutput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "zone_Id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "$ref": "#/components/schemas/TileType"
          },
          "x_Position": {
            "type": "integer",
            "format": "int32"
          },
          "y_Position": {
            "type": "integer",
            "format": "int32"
          },
          "updated_At": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "TileType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
...
    }
  }
}